+++ app/User.php
@@ -16,7 +16,7 @@ class User extends Authenticatable
      * @var array
      */
     protected $fillable = [
-        'first_name', 'last_name', 'promotion', 'email', 'password',
+        'first_name', 'last_name', 'promotion', 'email', 'password', 'role_id'
     ];
 
     /**