+++ app/Validators/PasswordFor.php
@@ -9,7 +9,7 @@ class PasswordFor
 {
     public function validate($attribute, $value, $parameters, $validator)
     {
-        $email = array_get($validator->getData(), $parameters[0], null);
+        $email = array_get($validator->getData(), $parameters[0]);
         $user = User::where('email', $email)->select('password')->first();
         if ($user !== null) {
             $password = $user->password;